[feat-31]: code refactoring
전체 파일 정리 및 내부 코드 정리
uuid 적용. 한 번에 대량의 request 시 Client마다 고유 id 지정
environment file 생성 event 함수 정의
.env 삭제, .bashrc에 환경변수 추가

Issues: #31
