--- add_data/schemas.py
@@ -1,46 +1,0 @@
-import inspect
-from typing import Type
-
-from fastapi import Form
-from pydantic import BaseModel
-from pydantic.fields import ModelField
-
-
-def as_form(cls: Type[BaseModel]):
-    new_parameters = []
-
-    for field_name, model_field in cls.__fields__.items():
-        model_field: ModelField
-
-        new_parameters.append(
-            inspect.Parameter(
-                model_field.alias,
-                inspect.Parameter.POSITIONAL_ONLY,
-                default=Form(...)
-                if not model_field.required
-                else Form(model_field.default),
-                annotation=model_field.outer_type_,
-            )
-        )
-
-    async def as_form_func(**data):
-        return cls(**data)
-
-    sig = inspect.signature(as_form_func)
-    sig = sig.replace(parameters=new_parameters)
-    as_form_func.__signature__ = sig
-    setattr(cls, "as_form", as_form_func)
-    return cls
-
-
-@as_form
-class HotelForm(BaseModel):
-    days_in_waiting_list: int = None
-    booking_changes: int = None
-    total_of_special_requests: int = None
-    lead_time: int = None
-    previous_cancellations: int = None
-    previous_bookings_not_canceled: int = None
-    market_segment: str = None
-    deposit_type: str = None
-    hotel: str = None
