--- add_data/static/css/styles.css
@@ -1,24 +1,0 @@
-.hantitle {
-    font-family: "Gothic A1";
-}
-
-fieldset.scheduler-border {
-  border: 1px groove #ddd !important;
-  padding: 0 1.4em 1.4em 1.4em !important;
-  margin: 1 0 1.5em 0 !important;
-  -webkit-box-shadow: 0px 0px 0px 0px #000;
-  box-shadow: 0px 0px 0px 0px #000;
-  margin-top: 30px !important;
-  width: 40%;
-}
-
-legend.scheduler-border {
-    font-size: 1.8em !important;
-    font-weight: bold !important;
-    text-align: left !important;
-    padding: 0 10px;
-    border-bottom: none;
-    margin-top: -15px;
-    background-color: white;
-    color: black;
-  }
