--- add_data/templates/result.html
@@ -1,59 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-  <title>예약 취소 여부 예측</title>
-{% endblock %}
-
-{% block body %}
-<div class="container-fluid p-5 bg-primary text-white text-center">
-  <h1 class="display-4 hantitle">예약 취소 여부 예측</h1>
-</div>
-<div class="container mt-5">
-  <div class="row">
-    <div class="col">
-      <table class="table">
-        <tr><td>호텔</td><td>{{form_data["hotel"]}}</td></tr>
-        <tr><td>예약 유형</td><td>
-          {% set ms = {"Online TA":"여행사(온라인)", "Offline TA/TO":"여행사(오프라인)/여행기획사", "Groups":"단체", "Direct":" 호텔 직접 예약", "Corporate":"회사","Complementary":"무상 제공 물품", "Aviation":"항공사", "Undefined":"기타"}  %}
-          {{ms[form_data["market_segment"]]}}</td></tr>
-        <tr><td>보증금 정책</td><td>
-          {% set repeated = {"No Deposit":"없음", "Non Refund":"환불 불가", "Refundable":"환불 가능"} %}
-          {{ repeated[form_data["deposit_type"]] }}</td></tr>
-        <tr><td>예약 기간</td><td>{{form_data["lead_time"]}}일</td></tr>
-        <tr><td>예약 확정까지 소요 일수</td><td>{{form_data["days_in_waiting_list"]}}일</td></tr>
-        <tr><td>취소 예약 횟수</td><td>{{form_data["previous_cancellations"]}}회</td></tr>
-        <tr><td>미취소 예약 횟수</td><td>{{form_data["previous_bookings_not_canceled"]}}회</td></tr>
-        <tr><td>예약 변경 횟수</td><td>{{form_data["booking_changes"]}}회</td></tr>
-        <tr><td>특별 요청 횟수</td><td>{{form_data["total_of_special_requests"]}}회</td></tr>
-      </table>
-    </div>
-    <div class="col">
-      <fieldset class="scheduler-border w-100">
-        <legend class="scheduler-border">예측 결과</legend>
-        <table class="w-100">
-          <tr>
-            <td class="w-50"><h4 class="mb-4">예약 취소 가능성</h4></td><td><h4 class="mb-4">예약 취소 확률</h4></td>
-          </tr>
-          <tr>
-            <td class="w-50">{% if result["predict"] == 1 %}
-              <p class="container-sm ms-4 h2 text-danger">높음</p>
-            {% else %}
-              <p class="container-sm ms-4 h3 text-success">낮음</p>
-            {% endif %}</td>
-            <td><p class="container-sm ms-4 h3 text-info">{{ result["proba"] }}%</p></td>
-          </tr>
-        </table>
-      </fieldset>
-      <fieldset class="scheduler-border w-100 mt-4">
-          <legend class="scheduler-border">성능 지표</legend>
-          <table class="w-100">
-            <tr><td class="w-50 text-secondary"><h4 class="text-secondary">auc</h4></td><td><h4 class="text-secondary">{{metrics["auc"] | round(4)}}</h4></td></tr>
-            <tr><td class="w-50 text-secondary"><h4 class="text-secondary">재현율(recall)</h4></td><td><h4 class="text-secondary">{{metrics["test recall"] | round(4)}}</h4></td></tr>
-            <tr><td class="w-50 text-secondary"><h4 class="text-secondary">정밀도(precision)</h4></td><td><h4 class="text-secondary">{{metrics["test precision"] | round(4)}}</h4></td></tr>
-            <tr><td class="w-50 text-secondary"><h4 class="text-secondary">정확도(accuracy)</h4></td><td><h4 class="text-secondary">{{metrics["test accuracy"] | round(4)}}</h4></td></tr>
-          </table>
-      </fieldset>
-    </div>
-  </div>
-</div>
-{% endblock %}
