+++ main.py
@@ -1,0 +1,26 @@
+import uvicorn
+from fastapi import FastAPI, Request, Depends
+from fastapi.responses import HTMLResponse
+from fastapi.staticfiles import StaticFiles
+from starlette.templating import Jinja2Templates
+import pandas as pd
+import sys, os
+
+sys.path.append(os.pardir)
+import routers
+
+app = FastAPI()
+templates = Jinja2Templates(directory="templates")
+app.mount("/static", StaticFiles(directory="static"), name="static")
+
+
+@app.get("/", response_class=HTMLResponse)
+async def root(request: Request):
+    return templates.TemplateResponse("index.html", {"request": request})
+
+
+# router
+app.include_router(routers.router)
+
+if __name__ == "__main__":
+    uvicorn.run("main:app", port=8003, reload=True)
